+++ src/modules/scenarioeditor/ScenarioEventsPanel.vue
@@ -1,10 +1,13 @@
 <script setup lang="ts">
 import { injectStrict } from "@/utils";
-import { activeScenarioKey } from "@/components/injects";
+import { activeScenarioKey, timeModalKey } from "@/components/injects";
 import { computed } from "vue";
 import { NScenarioEvent } from "@/types/internalModels";
 import PanelHeading from "@/components/PanelHeading.vue";
 import { useTimeFormatStore } from "@/stores/timeFormatStore";
+import ScenarioEventDropdownMenu from "@/modules/scenarioeditor/ScenarioEventDropdownMenu.vue";
+import { ScenarioEventAction } from "@/types/constants";
+import BaseButton from "@/components/BaseButton.vue";
 
 interface Props {
   selectOnly?: boolean;
@@ -15,9 +18,9 @@ const emit = defineEmits(["event-click"]);
 
 const {
   store,
-  time: { goToScenarioEvent },
+  time: { goToScenarioEvent, deleteScenarioEvent, updateScenarioEvent, addScenarioEvent },
 } = injectStrict(activeScenarioKey);
-
+const { getModalTimestamp } = injectStrict(timeModalKey);
 const fmt = useTimeFormatStore();
 const events = computed(() => store.state.events.map((id) => store.state.eventMap[id]));
 const t = computed(() => store.state.currentTime);
@@ -26,6 +29,30 @@ function onEventClick(event: NScenarioEvent) {
   if (!props.selectOnly) goToScenarioEvent(event);
   emit("event-click", event);
 }
+
+async function onAction(action: ScenarioEventAction, eventId: string) {
+  const scenarioEvent = store.state.eventMap[eventId];
+  if (!scenarioEvent) return;
+  switch (action) {
+    case "changeTime":
+      const newTimestamp = await getModalTimestamp(scenarioEvent.startTime, {
+        timeZone: store.state.info.timeZone,
+        title: "Set scenario event time",
+      });
+      if (newTimestamp !== undefined) {
+        updateScenarioEvent(eventId, { startTime: newTimestamp });
+      }
+      break;
+    case "delete":
+      deleteScenarioEvent(eventId);
+      break;
+  }
+}
+
+function addEvent() {
+  const day = new Date(t.value).getDate();
+  addScenarioEvent({ title: `Event ${day}`, startTime: t.value });
+}
 </script>
 <template>
   <div class="">
@@ -33,8 +60,8 @@ function onEventClick(event: NScenarioEvent) {
 
     <div class="flow-root">
       <ul class="mt-4">
-        <li v-for="(event, eventIdx) in events" :key="event.id">
-          <div class="relative pb-4">
+        <li v-for="(event, eventIdx) in events" :key="event.id" class="group flex">
+          <div class="relative flex-auto pb-4">
             <span
               v-if="eventIdx !== events.length - 1"
               class="absolute left-2 top-2 -ml-px h-full w-0.5 bg-gray-200"
@@ -64,8 +91,13 @@ function onEventClick(event: NScenarioEvent) {
               </div>
             </div>
           </div>
+          <div class="opacity-0 group-focus-within:opacity-100 group-hover:opacity-100">
+            <ScenarioEventDropdownMenu @action="onAction($event, event.id)" />
+          </div>
         </li>
       </ul>
     </div>
+
+    <BaseButton small @click="addEvent()" class="mt-4">Add scenario event</BaseButton>
   </div>
 </template>